refactor: use naming convention
